+++ components/bridge/impl/src/main/java/com/flipperdevices/bridge/impl/manager/service/FlipperVersionApiImpl.kt
@@ -89,6 +89,7 @@ class FlipperVersionApiImpl(
 
     override suspend fun reset(bleManager: UnsafeBleManager) {
         // Do nothing
+        semVerStateFlow.emit(null)
     }
 
     private fun onSupportedVersionReceived(bleManager: UnsafeBleManager, apiVersion: String) {