+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/viewmodel/FapHubCategoryViewModel.kt
@@ -5,12 +5,15 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.cachedIn
+import com.flipperdevices.core.pager.loadingPagingDataFlow
 import com.flipperdevices.faphub.category.impl.api.CATEGORY_OPEN_PATH_KEY
 import com.flipperdevices.faphub.dao.api.FapNetworkApi
 import com.flipperdevices.faphub.dao.api.model.FapCategory
 import com.flipperdevices.faphub.dao.api.model.SortType
+import com.flipperdevices.faphub.target.api.FlipperTargetProviderApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.launch
 import tangle.inject.TangleParam
@@ -19,15 +22,21 @@ import tangle.viewmodel.VMInject
 class FapHubCategoryViewModel @VMInject constructor(
     private val fapNetworkApi: FapNetworkApi,
     @TangleParam(CATEGORY_OPEN_PATH_KEY)
-    private val category: FapCategory
+    private val category: FapCategory,
+    targetProviderApi: FlipperTargetProviderApi
 ) : ViewModel() {
     private val sortTypeFlow = MutableStateFlow(SortType.UPDATE_AT_DESC)
 
-    val faps = sortTypeFlow.flatMapLatest { sortType ->
-        Pager(PagingConfig(pageSize = FAPS_PAGE_SIZE)) {
-            FapsCategoryPagingSource(fapNetworkApi, category, sortType)
+    val faps = combine(sortTypeFlow, targetProviderApi.getFlipperTarget()) { sortType, target ->
+        if (target == null) {
+            return@combine loadingPagingDataFlow()
+        }
+        return@combine Pager(
+            PagingConfig(pageSize = FAPS_PAGE_SIZE)
+        ) {
+            FapsCategoryPagingSource(fapNetworkApi, category, sortType, target)
         }.flow
-    }.cachedIn(viewModelScope)
+    }.flatMapLatest { it }.cachedIn(viewModelScope)
 
     fun getSortTypeFlow(): StateFlow<SortType> = sortTypeFlow