+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/retrofit/api/KtorfitCategoryApi.kt
@@ -2,12 +2,11 @@ package com.flipperdevices.faphub.dao.network.retrofit.api
 
 import com.flipperdevices.faphub.dao.network.retrofit.model.KtorfitCategory
 import de.jensklingenberg.ktorfit.http.GET
-import de.jensklingenberg.ktorfit.http.Path
+import de.jensklingenberg.ktorfit.http.Query
 
 interface KtorfitCategoryApi {
-    @GET("category/{uid}")
-    suspend fun get(@Path("uid") id: String): KtorfitCategory
-
     @GET("category")
-    suspend fun getAll(): List<KtorfitCategory>
+    suspend fun getAll(
+        @Query("api") sdkApi: String?
+    ): List<KtorfitCategory>
 }