+++ components/faphub/target/api/src/main/java/com/flipperdevices/faphub/target/api/FlipperTargetProviderApi.kt
@@ -4,6 +4,5 @@ import com.flipperdevices.faphub.target.model.FlipperTarget
 import kotlinx.coroutines.flow.StateFlow
 
 interface FlipperTargetProviderApi {
-    fun getFlipperTarget(): StateFlow<FlipperTarget>
-    suspend fun getFlipperTargetSync(): Result<FlipperTarget.Received>
+    fun getFlipperTarget(): StateFlow<FlipperTarget?>
 }