+++ components/faphub/target/api/src/main/java/com/flipperdevices/faphub/target/model/FlipperTarget.kt
@@ -3,10 +3,28 @@ package com.flipperdevices.faphub.target.model
 import com.flipperdevices.core.data.SemVer
 
 sealed class FlipperTarget {
-    object Retrieving : FlipperTarget()
     object Unsupported : FlipperTarget()
+
+    object NotConnected : FlipperTarget()
+
     data class Received(
         val target: String,
         val sdk: SemVer
     ) : FlipperTarget()
+
+    fun getTargetForServer(): String? {
+        return when (this) {
+            is Received -> target
+            Unsupported,
+            NotConnected -> null
+        }
+    }
+
+    fun getApiForServer(): String? {
+        return when (this) {
+            is Received -> sdk.toString()
+            Unsupported,
+            NotConnected -> null
+        }
+    }
 }