+++ components/faphub/target/impl/src/main/java/com/flipperdevices/faphub/target/impl/api/FlipperTargetProviderApiImpl.kt
@@ -14,7 +14,8 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -28,7 +29,7 @@ class FlipperTargetProviderApiImpl @Inject constructor(
     override val TAG = "FlipperTargetProviderApi"
 
     private val scope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
-    private val targetFlow = MutableStateFlow<FlipperTarget>(FlipperTarget.Retrieving)
+    private val targetFlow = MutableStateFlow<FlipperTarget?>(null)
 
     init {
         scope.launch {
@@ -38,22 +39,17 @@ class FlipperTargetProviderApiImpl @Inject constructor(
 
     override fun getFlipperTarget() = targetFlow.asStateFlow()
 
-    override suspend fun getFlipperTargetSync(): Result<FlipperTarget.Received> = runCatching {
-        return@runCatching when (
-            val receivedTarget = getFlipperTarget().first {
-                it != FlipperTarget.Retrieving
-            }
-        ) {
-            is FlipperTarget.Received -> receivedTarget
-            FlipperTarget.Retrieving -> error("Please, wait until target is ready")
-            FlipperTarget.Unsupported -> error("Fap catalog unsupported on this api version")
-        }
-    }
-
     private suspend fun subscribe() {
         info { "Start subscribe" }
         val serviceApi = serviceProvider.getServiceApi()
-        serviceApi.flipperVersionApi.getVersionInformationFlow().collect { version ->
+        combine(
+            serviceApi.connectionInformationApi.getConnectionStateFlow(),
+            serviceApi.flipperVersionApi.getVersionInformationFlow()
+        ) { connectionState, version ->
+            if (!connectionState.isConnected) {
+                targetFlow.emit(FlipperTarget.NotConnected)
+                return@combine
+            }
             info { "Receive version $version" }
             val sdkVersion = fetcher.getSdkApi(serviceApi.requestApi, version)
             info { "Sdk version is $sdkVersion" }
@@ -61,7 +57,7 @@ class FlipperTargetProviderApiImpl @Inject constructor(
                 FlipperSdkVersion.Unsupported,
                 FlipperSdkVersion.Error -> FlipperTarget.Unsupported
 
-                FlipperSdkVersion.InProgress -> FlipperTarget.Retrieving
+                FlipperSdkVersion.InProgress -> null
                 is FlipperSdkVersion.Received -> FlipperTarget.Received(
                     target = "f7",
                     sdk = sdkVersion.sdk
@@ -69,6 +65,6 @@ class FlipperTargetProviderApiImpl @Inject constructor(
             }
             info { "New target state is $newTargetState" }
             targetFlow.emit(newTargetState)
-        }
+        }.collect()
     }
 }