+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [FIX] Add confirm delete dialog and some small fixes
 - [Feature] FapHub MVPv0
 - [Feature] Update README with Fdroid badge
+- [Feature] Migrate faphub to target flow
 - [Feature] Update Geminio for create modules - simple, navigation, api/impl
 - [Refactor] Migrate all feature components modules from KSP to Anvil
 - [Refactor] Migrate to Ktorfit